requirements.txt
alembic==1.17.2
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==2.3.3
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.2.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
pillow==12.0.0
psycopg2-binary==2.9.11  # ← SOLO ESTE, elimina la línea psycopg2
python-dotenv==1.0.0
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==2.3.7
WTForms==3.0.1
lxml==6.0.2
PyPDF2==3.0.1
python-docx==1.2.0
pywin32==311
